requirements.txt
pandas>=1.5.0
reportlab>=4.0.0
openpyxl>=3.0.0
psycopg2-binary>=2.9.0
celery[redis]>=5.3.0
redis>=5.0.0
flower>=2.0.1
requests>=2.31.0
